deleted sheetManager file
