+++ index.php
@@ -1,9 +1,8 @@
-<?php 
+<?php
 require "config/global.php";
 require "model/db.php";
 require "model/categoriesManagement.php";
 require "model/sheetManagement.php";
-require "model/sheetManager.php";
 require "model/usersManagement.php";
 require "service/formChecker.php";
 require "service/router.php";
